@@ -1,35 +1,41 @@
 {
-  "name": "functions",
-  "description": "Cloud Functions for Firebase",
+  "name": "euphonia-functions",
   "scripts": {
+    "lint": "eslint --ext .js,.ts .",
     "build": "tsc",
-    "serve": "firebase emulators:start --only functions",
-    "shell": "firebase functions:shell",
+    "build:watch": "tsc --watch",
+    "serve": "npm run build && firebase emulators:start --only functions",
+    "shell": "npm run build && firebase functions:shell",
     "start": "npm run shell",
     "deploy": "firebase deploy --only functions",
     "logs": "firebase functions:log"
   },
   "engines": {
-    "node": "16"
+    "node": "22"
   },
   "main": "lib/functions/src/audioapp.js",
   "dependencies": {
-    "@google-cloud/firestore": "4.0.0",
-    "@google-cloud/storage": "5.18.2",
-    "body-parser": "1.19.2",
-    "cookie-parser": "1.4.6",
-    "express": "4.17.3",
-    "firebase-admin": "9.0.0",
-    "firebase-functions": "3.19.0",
-    "source-map-support": "0.5.21",
-    "typescript": "4.6.2",
-    "useragent": "2.3.0"
+    "@types/cookie-parser": "^1.4.8",
+    "@types/useragent": "^2.3.4",
+    "body-parser": "^1.20.3",
+    "cookie-parser": "^1.4.7",
+    "express": "^4.21.2",
+    "firebase": "^11.5.0",
+    "firebase-admin": "^13.2.0",
+    "firebase-functions": "^6.3.2",
+    "useragent": "^2.3.0"
   },
   "devDependencies": {
-    "@types/supertest": "2.0.12",
-    "firebase-functions-test": "0.2.0",
-    "intern": "4.10.1",
-    "supertest": "6.2.3"
+    "@types/supertest": "^6.0.2",
+    "@typescript-eslint/eslint-plugin": "^5.12.0",
+    "@typescript-eslint/parser": "^5.12.0",
+    "eslint": "^8.9.0",
+    "eslint-config-google": "^0.14.0",
+    "eslint-plugin-import": "^2.25.4",
+    "firebase-functions-test": "^3.4.1",
+    "intern": "^4.10.1",
+    "supertest": "^7.1.0",
+    "typescript": "^4.9.0"
   },
   "private": true
 }
