@@ -1,17 +1,32 @@
 {
-  "dependencies": {
-    "@google-cloud/storage": "5.18.2",
-    "@types/jquery": "3.5.14",
-    "cookie-parser": "1.4.6",
-    "express": "4.17.3",
-    "firebase": "9.6.8",
-    "firebase-admin": "9.0.0",
-    "firebase-functions": "3.19.0",
-    "firebase-tools": "11.29.0",
-    "multer": "1.4.4",
-    "typescript": "4.6.2"
-  },
+  "name": "euphonia",
   "scripts": {
-    "build": "tsc"
+    "dev:client": "esbuild src/app.ts src/admin/admin.ts --bundle --sourcemap=inline --outdir=public --watch",
+    "dev:functions": "cd functions && npm run dev",
+    "dev": "npm run gentemplate local && concurrently \"npm run dev:client\" \"npm run dev:functions\"",
+    "build:client": "esbuild src/app.ts src/admin/admin.ts --bundle --outdir=public --minify",
+    "build:functions": "cd functions && npm run build",
+    "build": "npm run gentemplate prod && npm run build:client && npm run build:functions",
+    "serve": "firebase serve --host localhost --port 8991",
+    "gentemplate": "node deployvars.js gentemplate",
+    "printvar": "node deployvars.js printvar",
+    "clean": "rm -rf public/app.js public/admin/admin.js functions/lib functions/src/firebaseconfig.ts src/firebaseconfig.ts firebase.json .firebaserc",
+    "deploy": "deploy/deploy.sh",
+    "postinstall": "cd functions && npm install"
+  },
+  "devDependencies": {
+    "@types/jquery": "^3.5.32",
+    "@typescript-eslint/eslint-plugin": "^5.12.0",
+    "@typescript-eslint/parser": "^5.12.0",
+    "eslint": "^8.9.0",
+    "eslint-config-google": "^0.14.0",
+    "eslint-plugin-import": "^2.25.4",
+    "concurrently": "^9.1.2",
+    "esbuild": "^0.25.1",
+    "firebase-tools": "^13.35.1",
+    "typescript": "^5.8.2"
+  },
+  "engines": {
+    "node": "22"
   }
 }
