@@ -15,92 +15,168 @@
  */
 
 // DAO object for working with stored data in GCS and Firestore
-import { Storage, File } from '@google-cloud/storage';
-import { Firestore, FieldPath, CollectionReference } from '@google-cloud/firestore';
-import * as firebaseconfig from './firebaseconfig';
-import * as schema from '../../commonsrc/schema';
-import { normalizeEmail, NotFoundError, ParamError, requireLanguage } from './util';
-import { clone, shuffle, toBatches } from '../../commonsrc/util';
+import { getApp } from "firebase/app";
+import {
+  getFirestore,
+  collection,
+  runTransaction,
+  query,
+  where,
+  limit,
+  orderBy,
+  doc,
+  getDoc,
+  getDocs,
+  documentId,
+} from "firebase/firestore";
+import { getStorage } from "firebase/storage";
+import * as firebaseconfig from "./firebaseconfig";
+import * as schema from "../../commonsrc/schema";
+import {
+  normalizeEmail,
+  NotFoundError,
+  ParamError,
+  requireLanguage,
+} from "./util";
+import { clone, shuffle, toBatches } from "../../commonsrc/util";
 
-type Transaction = FirebaseFirestore.Transaction;
+import type {
+  CollectionReference,
+  Firestore,
+  Transaction,
+} from "firebase/firestore";
+import type {
+  EUserData,
+  NewUserInfo,
+  EUserInfo,
+  EUserTaskInfo,
+  EUserTaskData,
+  EDeviceInfo,
+  ERecordingMetadata,
+  ERecordingData,
+  ETaskInfo,
+  ETaskData,
+  ETaskSetData,
+  ETaskSetInfo,
+  EAssignmentRule,
+  EConsentData,
+  EConsentInfo,
+  EConsentVersion,
+  EAgreementInfo,
+  TaskType,
+} from "../../commonsrc/schema";
 
 export class EStorage {
   static EUID_TRIES = 50;
 
-  firestore: Firestore;  // Firestore's TS configs seem broken
+  firestore: Firestore;
   storage: Storage;
-  
+
   constructor(deps?: any) {
-    this.firestore = deps ? deps.firestore : new Firestore();
-    this.storage = deps ? deps.storage : new Storage();
-    this.firestore.settings({ ignoreUndefinedProperties: true });
+    this.firestore = deps ? deps.firestore : getFirestore(getApp());
+    this.storage = deps ? deps.storage : getStorage(getApp());
+    // this.firestore.settings({ ignoreUndefinedProperties: true });
   }
 
   // Accessors for collections and sub-collections
 
   getUsersCollection(): CollectionReference {
-    return this.firestore.collection(`${schema.USERS_TABLE}`);
+    return collection(this.firestore, `${schema.USERS_TABLE}`);
   }
 
   getRecordingsSubcollection(euid: string): CollectionReference {
-    return this.firestore.collection(`${schema.RECORDINGS_TABLE}/${euid}/${schema.RECORDINGS_SUBCOLLECTION}`);
+    return collection(
+      this.firestore,
+      `${schema.RECORDINGS_TABLE}/${euid}/${schema.RECORDINGS_SUBCOLLECTION}`,
+    );
   }
 
   getUserTasksSubcollection(euid: string): CollectionReference {
-    return this.firestore.collection(`${schema.USERS_TABLE}/${euid}/${schema.TASKS_SUBCOLLECTION}`);
+    return collection(
+      this.firestore,
+      `${schema.USERS_TABLE}/${euid}/${schema.TASKS_SUBCOLLECTION}`,
+    );
   }
 
   getTaskSetsCollection(): CollectionReference {
-    return this.firestore.collection(`${schema.TASKSETS_TABLE}`);
+    return collection(this.firestore, `${schema.TASKSETS_TABLE}`);
   }
 
   getTasksSubcollection(tsid: string): CollectionReference {
-    return this.firestore.collection(`${schema.TASKSETS_TABLE}/${tsid}/${schema.TASKS_SUBCOLLECTION}`);
+    return collection(
+      this.firestore,
+      `${schema.TASKSETS_TABLE}/${tsid}/${schema.TASKS_SUBCOLLECTION}`,
+    );
   }
 
   getConsentsCollection(): CollectionReference {
-    return this.firestore.collection(`${schema.CONSENTS_TABLE}`);
+    return collection(this.firestore, `${schema.CONSENTS_TABLE}`);
   }
 
   // Runs a transaction with the given function.
   async run<X>(fn: (txn: Transaction) => Promise<X>): Promise<X> {
-    return this.firestore.runTransaction(fn);
+    return runTransaction(this.firestore, fn);
   }
 
   // Returns the EUser with the given Firebase identity ID, or null if not created.
-  async loadUserByFBUID(fbuid: string, opt_txn?: Transaction): Promise<EUser|null> {
-    const q = this.getUsersCollection().where('fbuid', '==', fbuid).limit(1);
-    const result = opt_txn ? await opt_txn.get(q) : await q.get();
-    if (result.size === 1) {
-      for (const doc of result.docs) {
-        const data = doc.data() as schema.EUserData;
+  async loadUserByFBUID(
+    fbuid: string,
+    opt_txn?: Transaction,
+  ): Promise<EUser | null> {
+    // const result = opt_txn ? await opt_txn.get(q) : await getDocs(q);
+
+    const q = await getDocs(
+      query(this.getUsersCollection(), where("fbuid", "==", fbuid), limit(1)),
+    );
+
+    if (q.size === 1) {
+      for (const doc of q.docs) {
+        const data = (
+          opt_txn ? await opt_txn.get(doc.ref) : doc.data()
+        ) as EUserData;
         return new EUser(this, schema.userPath(data.euid), data);
       }
     }
+
     return null;
   }
 
   // Returns the EUser with the given email, or null if not created.
-  async loadUserByEmail(email: string, opt_txn?: Transaction): Promise<EUser|null> {
+  async loadUserByEmail(
+    email: string,
+    opt_txn?: Transaction,
+  ): Promise<EUser | null> {
     const normedEmail = normalizeEmail(email);
-    const q = this.getUsersCollection().where('normalizedEmail', '==', normedEmail).limit(1);
-    const result = opt_txn ? await opt_txn.get(q) : await q.get();
+    const q = query(
+      this.getUsersCollection(),
+      where("normalizedEmail", "==", normedEmail),
+      limit(1),
+    );
+
+    // const result = opt_txn ? await opt_txn.get(q) : await q.get();
+    const result = await getDocs(q);
     if (result.size === 1) {
       for (const doc of result.docs) {
-        const data = doc.data() as schema.EUserData;
+        const data = doc.data() as EUserData;
         return new EUser(this, schema.userPath(data.euid), data);
       }
     }
+
     return null;
   }
 
   // Returns the EUser with the given EUID, or fails if not created.
   async loadUser(euid: string, opt_txn?: Transaction): Promise<EUser> {
     const path = schema.userPath(euid);
-    const userDoc = this.firestore.doc(path);
-    const existingUserData = opt_txn ? await opt_txn.get(userDoc) : await userDoc.get();
-    if (existingUserData.exists) {
-      return new EUser(this, path, existingUserData.data() as schema.EUserData);
+    const userDoc = doc(this.firestore, path);
+
+    // const existingUserData = opt_txn
+    //   ? await opt_txn.get(userDoc)
+    //   : await getDoc(userDoc);
+
+    const existingUserData = await getDoc(userDoc);
+    if (existingUserData.exists()) {
+      return new EUser(this, path, existingUserData.data() as EUserData);
     } else {
       throw new NotFoundError(`No such user: ${euid}`);
     }
@@ -108,48 +184,52 @@ export class EStorage {
 
   // Loads all users.
   async listUsers(): Promise<EUser[]> {
-    const data = await this.getUsersCollection().get();
-    return data.docs.map(doc => new EUser(this, doc.ref.path, doc.data() as schema.EUserData));
+    const data = await getDocs(this.getUsersCollection());
+
+    return data.docs.map(
+      (doc) => new EUser(this, doc.ref.path, doc.data() as EUserData),
+    );
   }
 
   // Creates a user if they don't exist, or claims their existing user account otherwise.
-  async signUpUser(newuser: schema.NewUserInfo): Promise<EUser> {
+  async signUpUser(newuser: NewUserInfo): Promise<EUser> {
     // First see if they already have an account and we can just claim it
     const fbuid = newuser.fbuid;
     if (!fbuid) {
-      throw new ParamError('FBUser info required for signup');
+      throw new ParamError("FBUser info required for signup");
     }
-    const user = await this.run(async txn => {
+    const user = await this.run(async (txn) => {
       let existingUser = await this.loadUserByFBUID(fbuid, txn);
       if (!existingUser) {
         // The user may have been pre-enrolled by an administrator
         existingUser = await this.loadUserByEmail(newuser.email, txn);
         if (!existingUser) {
-          return null;  // This is a fresh signup, no existing record to update
+          return null; // This is a fresh signup, no existing record to update
         }
       }
 
       // Claim and update the existing record
       existingUser.fbuid = fbuid;
-      existingUser.info.email = newuser.email;  // this can be slightly different due to normalization
+      existingUser.info.email = newuser.email; // this can be slightly different due to normalization
       existingUser.info.fbname = newuser.fbname;
       existingUser.info.signupTimestamp = newuser.signupTimestamp;
       existingUser.info.demographics = newuser.demographics;
       existingUser.update(txn);
+
       return existingUser;
     });
 
     if (user != null) {
-      return user;  // Successfully claimed an existing user record
+      return user; // Successfully claimed an existing user record
     }
 
     // Otherwise, this is a fresh signup, create an all new user record
-    const [createdUser, ] = await this.createUser(newuser);
+    const [createdUser] = await this.createUser(newuser);
     return createdUser;
   }
 
   // Creates a new EUser and runs all enrollment rules; also returns affected task set counters
-  async createUser(newuser: schema.NewUserInfo): Promise<[EUser, ETaskSet[]]> {
+  async createUser(newuser: NewUserInfo): Promise<[EUser, ETaskSet[]]> {
     const taskSets = new Set<ETaskSet>();
     let user = await this.createUser_(newuser);
 
@@ -163,46 +243,51 @@ export class EStorage {
   }
 
   // Creates a new EUser, generating a new unique EUID for them within a transaction.
-  private async createUser_(newuser: schema.NewUserInfo): Promise<EUser> {
+  private async createUser_(newuser: NewUserInfo): Promise<EUser> {
     requireLanguage(newuser.language);
 
     for (let tries = 0; tries < EStorage.EUID_TRIES; tries++) {
       try {
-        return await this.run(async txn => {
+        return await this.run(async (txn) => {
           // Make sure the normalized email is unique
           const user = await this.loadUserByEmail(newuser.email, txn);
           if (user) {
-            throw new Error(`Email address already in use: ${user.info.email} (${user.normalizedEmail})`);
+            throw new Error(
+              `Email address already in use: ${user.info.email} (${user.normalizedEmail})`,
+            );
           } else {
             // Try to create a new user with a new EUID. We might have to retry
             return await this.createUserTxn(txn, newuser);
           }
         });
-
       } catch (e) {
-        if (`${e}`.indexOf('retry: EUID in use') !== -1) {
-          console.log('EUID collision, trying again...');
-          continue;  // EUID collection, try again
+        if (`${e}`.indexOf("retry: EUID in use") !== -1) {
+          console.log("EUID collision, trying again...");
+          continue; // EUID collection, try again
         }
         throw e;
       }
     }
-    throw new Error(`Could not generate EUID after ${EStorage.EUID_TRIES} tries`);
+    throw new Error(
+      `Could not generate EUID after ${EStorage.EUID_TRIES} tries`,
+    );
   }
 
   // One shot at creating a user within a transaction, after existence checks are done.
-  private async createUserTxn(txn: Transaction, newuser: schema.NewUserInfo) {
+  private async createUserTxn(txn: Transaction, newuser: NewUserInfo) {
     // Allocate a new EUID within a transaction and claim it for this new user.
-    const newEuid = 'E' + `${Math.floor(Math.random() * 99999)}`.padStart(5, '0');
+    const newEuid =
+      "E" + `${Math.floor(Math.random() * 99999)}`.padStart(5, "0");
     const path = schema.userPath(newEuid);
-    const doc = this.firestore.doc(path);
-    const result = await txn.get(doc);
-    if (result.exists) {
+    const docRef = doc(this.firestore, path);
+
+    const result = await txn.get(docRef);
+    if (result.exists()) {
       throw new Error(`retry: EUID in use: ${newEuid}`);
     }
 
     // This EUID is unused, create a full user record
-    const info: schema.EUserInfo = {
+    const info: EUserInfo = {
       euid: newEuid,
       email: newuser.email,
       name: newuser.name,
@@ -218,23 +303,31 @@ export class EStorage {
       lastRecordingTimestamp: 0,
       numTasks: 0,
       numCompletedTasks: 0,
-      numAssignmentsByTaskSet: []
+      numAssignmentsByTaskSet: [],
     };
 
-    const fsdata: schema.EUserData = {
+    const fsdata: EUserData = {
       euid: newEuid,
-      fbuid: newuser.fbuid,  // this will be unset for admin-created users
+      fbuid: newuser.fbuid, // this will be unset for admin-created users
       normalizedEmail: normalizeEmail(newuser.email),
-      info: JSON.stringify(info)
+      info: JSON.stringify(info),
     };
 
-    txn.set(doc, fsdata);
+    txn.set(docRef, fsdata);
+
     return new EUser(this, path, fsdata);
   }
 
   // Writes audio to GCS for the given user, and creates a JSON as well as a Firestore record.
-  async createRecording(user: EUser, task: schema.EUserTaskInfo, localDate: string,
-      tzOffset: number, mimeType: string, deviceInfo: schema.EDeviceInfo, audioData: Buffer): Promise<[EUser, EUserTask, ERecording]> {
+  async createRecording(
+    user: EUser,
+    task: EUserTaskInfo,
+    localDate: string,
+    tzOffset: number,
+    mimeType: string,
+    deviceInfo: EDeviceInfo,
+    audioData: Buffer,
+  ): Promise<[EUser, EUserTask, ERecording]> {
     const euid = user.euid;
     const taskId = task.id;
     const timestamp = Date.now();
@@ -243,10 +336,10 @@ export class EStorage {
     // Load the taskset outside of the transaction so we don't contend on it
     const taskSet = await this.requireTaskSet(task.taskSetId);
 
-    const metadata: schema.ERecordingMetadata = {
+    const metadata: ERecordingMetadata = {
       euid,
       name: basename,
-      platform: 'audiotoolv1',
+      platform: "audiotoolv1",
       project: taskSet.info.id,
       task: task.id,
       language: taskSet.info.language,
@@ -258,10 +351,10 @@ export class EStorage {
       consents: user.info.consents,
       fileSize: audioData.length,
       mimeType,
-      deviceInfo
+      deviceInfo,
     };
 
-    if (task.task.taskType === 'prompt') {
+    if (task.task.taskType === "prompt") {
       // For verbatim repeat tasks, we assume that the user read the prompt exactly until found otherwise
       metadata.transcript = task.task.prompt;
     }
@@ -270,10 +363,12 @@ export class EStorage {
     const bucket = this.storage.bucket(firebaseconfig.BUCKET);
     const dirname = `${firebaseconfig.RECORDING_PATH}/${euid}`;
     await bucket.file(`${dirname}/${basename}.wav`).save(audioData);
-    await bucket.file(`${dirname}/${basename}.json`).save(JSON.stringify(metadata, null, 2));
+    await bucket
+      .file(`${dirname}/${basename}.json`)
+      .save(JSON.stringify(metadata, null, 2));
 
     // Also put a record in Firestore so the GUI can easily query, with updated user and task counters.
-    return await this.firestore.runTransaction(async txn => {
+    return await runTransaction(this.firestore, async (txn: Transaction) => {
       // Reload the user and task within the transaction to get the freshest versions.
       const user = await this.loadUser(euid, txn);
       const usertask = await user.loadTask(txn, taskId);
@@ -281,9 +376,9 @@ export class EStorage {
 
       // Save the recording
       const path = schema.recordingPath(euid, timestamp);
-      const recordingDoc = this.firestore.doc(path);
-      const fsdata: schema.ERecordingData = {
-        metadata: JSON.stringify(metadata)
+      const recordingDoc = doc(this.firestore, path);
+      const fsdata: ERecordingData = {
+        metadata: JSON.stringify(metadata),
       };
       txn.set(recordingDoc, fsdata);
 
@@ -294,10 +389,10 @@ export class EStorage {
       usertask.update(txn);
 
       // Update the user's recording and task counters
-      user.info.numRecordings = await this.countUserRecordings(user.euid) + 1;
+      user.info.numRecordings = (await this.countUserRecordings(user.euid)) + 1;
       user.info.lastRecordingTimestamp = timestamp;
       if (prevTimestamp === 0) {
-        user.info.numCompletedTasks += 1;  // Don't do this for re-records
+        user.info.numCompletedTasks += 1; // Don't do this for re-records
       }
       user.update(txn);
 
@@ -311,31 +406,41 @@ export class EStorage {
   }
 
   // Returns the two GCS File objects for a recording and its metadata, throws an error if not found.
-  findRecordingFiles(euid: string, basename: string): [File, File] {
+  findRecordingFiles(euid: string, basename: string) {
     const bucket = this.storage.bucket(firebaseconfig.BUCKET);
     const wpath = `${firebaseconfig.RECORDING_PATH}/${euid}/${basename}.wav`;
     const jpath = `${firebaseconfig.RECORDING_PATH}/${euid}/${basename}.json`;
+
     const wfile = bucket.file(wpath);
     const jfile = bucket.file(jpath);
+
     if (!wfile || !jfile) {
       throw new NotFoundError(`No such recording: ${basename}`);
     }
+
     return [wfile, jfile];
   }
 
   // Queries the user's recordings to count how many there are.
   private async countUserRecordings(euid: string): Promise<number> {
-    const data = await this.getRecordingsSubcollection(euid).get();
+    const data = await getDocs(this.getRecordingsSubcollection(euid));
     return data.size;
   }
 
   // Loads a taskSet by ID, which should be unique.
-  async loadTaskSet(id: string, opt_txn?: Transaction): Promise<ETaskSet|undefined> {
+  async loadTaskSet(
+    id: string,
+    opt_txn?: Transaction,
+  ): Promise<ETaskSet | undefined> {
     const path = schema.taskSetPath(id);
-    const doc = this.firestore.doc(path);
-    const existingData = opt_txn ? await opt_txn.get(doc) : await doc.get();
-    if (existingData.exists) {
-      return new ETaskSet(this, path, existingData.data() as schema.ETaskSetData);
+    const docRef = doc(this.firestore, path);
+
+    const existingData = opt_txn
+      ? await opt_txn.get(docRef)
+      : await getDoc(docRef);
+
+    if (existingData.exists()) {
+      return new ETaskSet(this, path, existingData.data() as ETaskSetData);
     } else {
       return undefined;
     }
@@ -351,45 +456,52 @@ export class EStorage {
   }
 
   // Loads all TaskSets.
-  async listTaskSets() : Promise<ETaskSet[]> {
-    const data = await this.getTaskSetsCollection().get();
-    return data.docs.map(doc => new ETaskSet(this, doc.ref.path, doc.data() as schema.ETaskSetData));
+  async listTaskSets(): Promise<ETaskSet[]> {
+    const data = await getDocs(this.getTaskSetsCollection());
+    return data.docs.map(
+      (doc) => new ETaskSet(this, doc.ref.path, doc.data() as ETaskSetData),
+    );
   }
 
   // Creates a new TaskSet
   async createTaskSet(id: string, name: string, language: string) {
-    return await this.run(async txn => {
+    return await this.run(async (txn) => {
       let taskSet = await this.loadTaskSet(id, txn);
       if (taskSet) {
         throw new Error(`TaskSet with this id already exists: ${id}`);
       }
       const path = schema.taskSetPath(id);
-      const doc = this.firestore.doc(path);
-      const info: schema.ETaskSetInfo = {
+      const docRef = doc(this.firestore, path);
+
+      const info: ETaskSetInfo = {
         id,
         name,
         creationTimestamp: Date.now(),
         language,
         rules: [],
         numAssignedTasks: 0,
-        numAssignedUsers: 0
+        numAssignedUsers: 0,
       };
-      const fsdata: schema.ETaskSetData = {info: JSON.stringify(info)};
-      await txn.set(doc, fsdata);
+
+      const fsdata: ETaskSetData = { info: JSON.stringify(info) };
+      txn.set(docRef, fsdata);
+
       return new ETaskSet(this, path, fsdata);
     });
   }
 
   // Loads all of the Assignment Rules that match a particular user, returns [taskSetId, rule] tuples.
-  async loadAssignmentRules(user: EUser): Promise<[ETaskSet, schema.EAssignmentRule][]> {
+  async loadAssignmentRules(
+    user: EUser,
+  ): Promise<[ETaskSet, EAssignmentRule][]> {
     const userTags: Set<string> = new Set(user.info.tags);
     const userLanguage: string = user.info.language;
 
     // Rules have to match the language and one of the user's tags.
-    const result: [ETaskSet, schema.EAssignmentRule][] = [];
+    const result: [ETaskSet, EAssignmentRule][] = [];
     for (const taskSet of await this.listTaskSets()) {
       if (taskSet.info.language !== userLanguage) {
-        continue;  // This rule's language doesn't match the user
+        continue; // This rule's language doesn't match the user
       }
       for (const rule of taskSet.info.rules) {
         if (rule.tags.length > 0) {
@@ -401,25 +513,33 @@ export class EStorage {
             }
           }
           if (!tagMatch) {
-            continue;  // This rule has no tags in common with the user, skip it
+            continue; // This rule has no tags in common with the user, skip it
           }
         }
-        result.push([taskSet, rule]);  // This rule matches
+        result.push([taskSet, rule]); // This rule matches
       }
     }
 
     // Order the rules
-    result.sort((a, b) => a[1].order < b[1].order ? -1 : (a[1].order === b[1].order ? 0 : 1));
+    result.sort((a, b) =>
+      a[1].order < b[1].order ? -1 : a[1].order === b[1].order ? 0 : 1,
+    );
     return result;
   }
 
   // Loads a consent by ID, which should be unique.
-  async loadConsent(id: string, opt_txn?: Transaction): Promise<EConsent|undefined> {
+  async loadConsent(
+    id: string,
+    opt_txn?: Transaction,
+  ): Promise<EConsent | undefined> {
     const path = schema.consentPath(id);
-    const doc = this.firestore.doc(path);
-    const existingData = opt_txn ? await opt_txn.get(doc) : await doc.get();
-    if (existingData.exists) {
-      return new EConsent(this, path, existingData.data() as schema.EConsentData);
+    const docRef = doc(this.firestore, path);
+    const existingData = opt_txn
+      ? await opt_txn.get(docRef)
+      : await getDoc(docRef);
+
+    if (existingData.exists()) {
+      return new EConsent(this, path, existingData.data() as EConsentData);
     } else {
       return undefined;
     }
@@ -435,23 +555,29 @@ export class EStorage {
   }
 
   // Loads all Consents.
-  async listConsents() : Promise<EConsent[]> {
-    const data = await this.getConsentsCollection().get();
-    return data.docs.map(doc => new EConsent(this, doc.ref.path, doc.data() as schema.EConsentData));
+  async listConsents(): Promise<EConsent[]> {
+    const data = await getDocs(this.getConsentsCollection());
+    return data.docs.map(
+      (doc) => new EConsent(this, doc.ref.path, doc.data() as EConsentData),
+    );
   }
 
   // Same as above, but returns only the subset of consents that apply to a user with the given language and tags.
-  async listApplicableConsents(language: string, tags: string[], now: number) : Promise<EConsent[]> {
+  async listApplicableConsents(
+    language: string,
+    tags: string[],
+    now: number,
+  ): Promise<EConsent[]> {
     const result: Set<EConsent> = new Set();
     for (const consent of await this.listConsents()) {
       if (!consent.info.active) {
-        continue;  // suppress this inactive consent
+        continue; // suppress this inactive consent
       }
       if (consent.info.language !== language) {
-        continue;  // not applicable to this user
+        continue; // not applicable to this user
       }
       if (!consent.getLiveVersion(now)) {
-        continue;  // this consent is post-dated into the future, not yet relevant
+        continue; // this consent is post-dated into the future, not yet relevant
       }
       if (consent.info.tags.length > 0) {
         // A tagged consent means it only applies to certain users
@@ -471,15 +597,22 @@ export class EStorage {
   }
 
   // Creates a new Consent
-  async createConsent(id: string, name: string, language: string, tags: string[], optional: boolean) {
-    return await this.run(async txn => {
+  async createConsent(
+    id: string,
+    name: string,
+    language: string,
+    tags: string[],
+    optional: boolean,
+  ) {
+    return await this.run(async (txn) => {
       let consent = await this.loadConsent(id, txn);
       if (consent) {
         throw new Error(`Consent with this id already exists: ${id}`);
       }
       const path = schema.consentPath(id);
-      const doc = this.firestore.doc(path);
-      const info: schema.EConsentInfo = {
+      const docRef = doc(this.firestore, path);
+
+      const info: EConsentInfo = {
         id,
         name,
         language,
@@ -487,23 +620,25 @@ export class EStorage {
         optional,
         active: true,
         creationTimestamp: Date.now(),
-        versions: []
+        versions: [],
       };
-      const fsdata: schema.EConsentData = {info: JSON.stringify(info)};
-      txn.set(doc, fsdata);
+
+      const fsdata: EConsentData = { info: JSON.stringify(info) };
+      txn.set(docRef, fsdata);
+
       return new EConsent(this, path, fsdata);
     });
   }
 
   // Returns the GCS file object for a given consent HTML
-  getConsentFile(consentId: string, version: number): File {
+  getConsentFile(consentId: string, version: number) {
     const bucket = this.storage.bucket(firebaseconfig.BUCKET);
     const dirname = `${firebaseconfig.CONSENT_PATH}`;
     return bucket.file(`${dirname}/${consentId}-${version}.html`);
   }
 
   // Returns the GCS file object for a graphical task
-  getImageFile(taskSetId: string, taskId: string): File {
+  getImageFile(taskSetId: string, taskId: string) {
     const bucket = this.storage.bucket(firebaseconfig.BUCKET);
     const dirname = `${firebaseconfig.IMAGETASKS_PATH}`;
     return bucket.file(`${dirname}/image_${taskSetId}_${taskId}.jpg`);
@@ -513,19 +648,19 @@ export class EStorage {
 // DAO for one logged in user of the system.
 export class EUser {
   parent: EStorage;
-  path: string;  // Firestore document path
+  path: string; // Firestore document path
 
-  euid: string;  // immutable ID
-  fbuid?: string;  // set when account is claimed
-  normalizedEmail: string;  // indexed version of email
-  info: schema.EUserInfo;
+  euid: string; // immutable ID
+  fbuid?: string; // set when account is claimed
+  normalizedEmail: string; // indexed version of email
+  info: EUserInfo;
 
-  constructor(parent: EStorage, path: string, fsdata: schema.EUserData) {
-    this.parent = parent;  // EStorage system
-    this.path = path;  // Firebase document reference path
+  constructor(parent: EStorage, path: string, fsdata: EUserData) {
+    this.parent = parent; // EStorage system
+    this.path = path; // Firebase document reference path
 
-    this.euid = fsdata.euid;  // Immutable
-    this.fbuid = fsdata.fbuid;  // This can be null for unclaimed accounts
+    this.euid = fsdata.euid; // Immutable
+    this.fbuid = fsdata.fbuid; // This can be null for unclaimed accounts
     this.normalizedEmail = fsdata.normalizedEmail;
 
     // Everything else from the JSON blob: email, name, language, counters, etc
@@ -550,7 +685,9 @@ export class EUser {
           t[1] += delta;
           this.info.numTasks += delta;
         } else {
-          console.log(`Warning: attempted to decrement counter below zero: ${taskSetId} for ${this.euid}`);
+          console.log(
+            `Warning: attempted to decrement counter below zero: ${taskSetId} for ${this.euid}`,
+          );
           this.info.numTasks -= t[1];
           t[1] = 0;
         }
@@ -564,8 +701,10 @@ export class EUser {
       this.info.numTasks += delta;
       return delta;
     } else {
-      console.log(`Warning: attempted to decrement unassigned taskset: ${taskSetId} for ${this.euid}`);
-      return 0;  // we don't allow counters to go below zero
+      console.log(
+        `Warning: attempted to decrement unassigned taskset: ${taskSetId} for ${this.euid}`,
+      );
+      return 0; // we don't allow counters to go below zero
     }
   }
 
@@ -574,23 +713,29 @@ export class EUser {
     const obj: any = {
       // EUID is immutable, never update it
       normalizedEmail: this.normalizedEmail,
-      info: JSON.stringify(this.info)
+      info: JSON.stringify(this.info),
     };
+
     if (this.fbuid) {
-      obj['fbuid'] = this.fbuid;  // Only set property keys with non-null values
+      obj["fbuid"] = this.fbuid; // Only set property keys with non-null values
     }
-    txn.update(this.parent.firestore.doc(this.path), obj);
+
+    txn.update(doc(this.parent.firestore, this.path), obj);
   }
 
   // Returns true if the user's agreements match the live versions of all applicable consents
   async isConsented(now: number): Promise<boolean> {
-    for (const c of await this.parent.listApplicableConsents(this.info.language, this.info.tags, now)) {
+    for (const c of await this.parent.listApplicableConsents(
+      this.info.language,
+      this.info.tags,
+      now,
+    )) {
       if (c.info.optional) {
-        continue;  // Permit user not to have consented if the consent is optional
+        continue; // Permit user not to have consented if the consent is optional
       }
       const v = c.getLiveVersion(now);
       if (!v || !this.isConsented_(c.info.id, v.version)) {
-        return false;  // The user doesn't have a current agreement record for this consent
+        return false; // The user doesn't have a current agreement record for this consent
       }
     }
     return true;
@@ -599,19 +744,29 @@ export class EUser {
   // Returns true if this user has an agreement to the given exact consent and version
   private isConsented_(consentId: string, version: number): boolean {
     for (const agreement of this.info.consents) {
-      if (agreement.consentId === consentId && agreement.version === version &&
-          agreement.consentTimestamp !== 0 && agreement.revokeTimestamp === 0) {
-        return true;  // Found a matching and non-revoked agreement
+      if (
+        agreement.consentId === consentId &&
+        agreement.version === version &&
+        agreement.consentTimestamp !== 0 &&
+        agreement.revokeTimestamp === 0
+      ) {
+        return true; // Found a matching and non-revoked agreement
       }
     }
     return false;
   }
 
   // Records that the user agreed to a particular consent.
-  async addConsent(now: number, agreement: schema.EAgreementInfo): Promise<EUser> {
-    return await this.parent.run(async txn => {
-      const consent = await this.parent.requireConsent(agreement.consentId, txn);
-      let [exact, others] = this.findConsents(agreement.consentId, agreement.version);
+  async addConsent(now: number, agreement: EAgreementInfo): Promise<EUser> {
+    return await this.parent.run(async (txn) => {
+      const consent = await this.parent.requireConsent(
+        agreement.consentId,
+        txn,
+      );
+      let [exact, others] = this.findConsents(
+        agreement.consentId,
+        agreement.version,
+      );
       if (exact && exact.revokeTimestamp === 0) {
         // Already consented, just update the timestamps. No change to counters.
         exact.consentTimestamp = now;
@@ -622,7 +777,7 @@ export class EUser {
       // Note the user's new consent and count it, or reinstate a previously revoked consent
       if (exact) {
         exact.consentTimestamp = now;
-        exact.revokeTimestamp = 0;  // reinstate
+        exact.revokeTimestamp = 0; // reinstate
       } else {
         agreement.consentTimestamp = now;
         agreement.revokeTimestamp = 0;
@@ -660,14 +815,19 @@ export class EUser {
   }
 
   // Returns the matching consent, plus any other consents of different versions with the same ID.
-  private findConsents(consentId: string, version: number): [schema.EAgreementInfo|undefined, schema.EAgreementInfo[]] {
-    let exact: schema.EAgreementInfo|undefined;  // ID and version match
-    const others: schema.EAgreementInfo[] = [];  // only IDs match
+  private findConsents(
+    consentId: string,
+    version: number,
+  ): [EAgreementInfo | undefined, EAgreementInfo[]] {
+    let exact: EAgreementInfo | undefined; // ID and version match
+    const others: EAgreementInfo[] = []; // only IDs match
     for (const c of this.info.consents) {
       if (c.consentId === consentId) {
         if (c.version === version) {
           if (exact) {
-            throw new Error(`User ${this.info.euid} has multiple exact consent matches: ${consentId}-${version}`);
+            throw new Error(
+              `User ${this.info.euid} has multiple exact consent matches: ${consentId}-${version}`,
+            );
           }
           exact = c;
         } else {
@@ -680,44 +840,71 @@ export class EUser {
 
   // Returns the user's tasks, if any
   async listTasks(): Promise<EUserTask[]> {
-    const data = await this.parent.getUserTasksSubcollection(this.euid).orderBy('order').get();
-    return data.docs.map(doc => new EUserTask(this, doc.ref.path, doc.data() as schema.EUserTaskData));
+    const data = await getDocs(
+      query(this.parent.getUserTasksSubcollection(this.euid), orderBy("order")),
+    );
+
+    return data.docs.map(
+      (doc) => new EUserTask(this, doc.ref.path, doc.data() as EUserTaskData),
+    );
   }
 
   // Returns the user's recordings, if any
   async listRecordings(): Promise<ERecording[]> {
-    const data = await this.parent.getRecordingsSubcollection(this.euid).get();
-    return data.docs.map(doc => new ERecording(this, doc.ref.path, doc.data() as schema.ERecordingData));
+    const data = await getDocs(
+      this.parent.getRecordingsSubcollection(this.euid),
+    );
+
+    return data.docs.map(
+      (doc) => new ERecording(this, doc.ref.path, doc.data() as ERecordingData),
+    );
   }
 
   // Loads one UserTask using a transaction, returns null if not found
   async loadTask(txn: Transaction, taskId: string): Promise<EUserTask> {
     const path = schema.userTaskPath(this.euid, taskId);
-    const existingTaskData = await txn.get(this.parent.firestore.doc(path));
-    if (existingTaskData.exists) {
-      return new EUserTask(this, path, existingTaskData.data() as schema.EUserTaskData);
+    const existingTaskData = await txn.get(doc(this.parent.firestore, path));
+
+    if (existingTaskData.exists()) {
+      return new EUserTask(
+        this,
+        path,
+        existingTaskData.data() as EUserTaskData,
+      );
     } else {
       throw new NotFoundError(`No such task: ${taskId}`);
     }
   }
 
   // Loads one Recording using a transaction. Throws an error if not found.
-  async loadRecording(txn: Transaction, timestamp: number): Promise<ERecording> {
+  async loadRecording(
+    txn: Transaction,
+    timestamp: number,
+  ): Promise<ERecording> {
     const path = schema.recordingPath(this.euid, timestamp);
-    const data = await txn.get(this.parent.firestore.doc(path));
-    if (data.exists) {
-      return new ERecording(this, path, data.data() as schema.ERecordingData);
+    const data = await txn.get(doc(this.parent.firestore, path));
+
+    if (data.exists()) {
+      return new ERecording(this, path, data.data() as ERecordingData);
     } else {
       throw new NotFoundError(`Recording not found: ${timestamp}`);
     }
   }
 
   // Adds the list of tasks to a user and updates the user and TaskSet counters. Max 500 tasks please.
-  async assignTasks(txn: Transaction, tasks: ETask[], taskSet: ETaskSet): Promise<void> {
+  async assignTasks(
+    txn: Transaction,
+    tasks: ETask[],
+    taskSet: ETaskSet,
+  ): Promise<void> {
     const tsid = taskSet.info.id;
     // Get the currently highest task order number
     const collection = this.parent.getUserTasksSubcollection(this.euid);
-    const q = await txn.get(collection.orderBy('order', 'desc').limit(1));
+    const q = await getDocs(
+      query(collection, orderBy("order", "desc"), limit(1)),
+    );
+
+    // const q = await txn.get(collection.orderBy("order", "desc").limit(1));
     let order = 0;
     for (const doc of q.docs) {
       order = doc.data().order;
@@ -728,17 +915,18 @@ export class EUser {
     // Save these new tasks, in shuffled order.
     const timestamp = Date.now();
     for (const task of shuffle([...tasks])) {
-      const newDoc = collection.doc();
-      const info: schema.EUserTaskInfo = {
+      const newDoc = doc(collection);
+      const info: EUserTaskInfo = {
         id: newDoc.id,
         taskSetId: tsid,
         task: task.info,
         assignedTimestamp: timestamp,
-        recordedTimestamp: 0
+        recordedTimestamp: 0,
       };
-      const fsdata: schema.EUserTaskData = {
+      const fsdata: EUserTaskData = {
         order: order++,
-        info: JSON.stringify(info)};
+        info: JSON.stringify(info),
+      };
       txn.set(newDoc, fsdata);
     }
 
@@ -747,14 +935,17 @@ export class EUser {
     const newNum = this.changeNumTasks(tsid, tasks.length);
     taskSet.info.numAssignedTasks += tasks.length;
     if (oldNum === 0 && newNum > 0) {
-      taskSet.info.numAssignedUsers++;  // this user didn't have tasks from this taskset before
+      taskSet.info.numAssignedUsers++; // this user didn't have tasks from this taskset before
     }
     taskSet.update(txn);
     this.update(txn);
   }
 
   // Removes the list of tasks from a user and updates the user's task counters.
-  async removeTasks(txn: Transaction, idTuples: [string, string][]): Promise<ETaskSet[]> {
+  async removeTasks(
+    txn: Transaction,
+    idTuples: [string, string][],
+  ): Promise<ETaskSet[]> {
     // First load all involved tasksets so we can update their counters.
     const tasksByParent = await this.loadTaskTuples(txn, idTuples);
 
@@ -765,14 +956,17 @@ export class EUser {
       const newNum = this.changeNumTasks(taskSet.info.id, -taskIds.length);
       if (oldNum > 0 && newNum <= 0) {
         // This user has no tasks left from this taskset, so they drop out of the user counter
-        taskSet.info.numAssignedUsers = Math.max(0, taskSet.info.numAssignedUsers - 1);
+        taskSet.info.numAssignedUsers = Math.max(
+          0,
+          taskSet.info.numAssignedUsers - 1,
+        );
       }
     }
 
     // Remove these tasks; the client was supposed to have already checked if they're unrecorded
     for (const [, taskId] of idTuples) {
       const taskpath = schema.userTaskPath(this.euid, taskId);
-      txn.delete(this.parent.firestore.doc(taskpath));
+      txn.delete(doc(this.parent.firestore, taskpath));
     }
 
     // Update the user's task counters
@@ -787,7 +981,10 @@ export class EUser {
   }
 
   // For each [taskSetId, taskId] tuple, loads the parent TaskSet and organizes all its taskIds in a map.
-  private async loadTaskTuples(txn: Transaction, idTuples: Array<[string, string]>): Promise<Map<ETaskSet, string[]>> {
+  private async loadTaskTuples(
+    txn: Transaction,
+    idTuples: Array<[string, string]>,
+  ): Promise<Map<ETaskSet, string[]>> {
     const taskSets = new Map();
     const result = new Map();
     for (const [taskSetId, taskId] of idTuples) {
@@ -806,11 +1003,11 @@ export class EUser {
 // DAO for one task assigned to a user
 export class EUserTask {
   parent: EUser;
-  path: string;  // Firestore document path
+  path: string; // Firestore document path
   order: number;
-  info: schema.EUserTaskInfo;
+  info: EUserTaskInfo;
 
-  constructor(parent: EUser, path: string, fsdata: schema.EUserTaskData) {
+  constructor(parent: EUser, path: string, fsdata: EUserTaskData) {
     this.parent = parent;
     this.path = path;
     this.order = fsdata.order;
@@ -819,9 +1016,9 @@ export class EUserTask {
 
   // Updates this UserTask with any changes carried by this DAO.
   update(txn: Transaction) {
-    txn.update(this.parent.parent.firestore.doc(this.path), {
+    txn.update(doc(this.parent.parent.firestore, this.path), {
       // Never mutate order
-      info: JSON.stringify(this.info)
+      info: JSON.stringify(this.info),
     });
   }
 }
@@ -829,22 +1026,38 @@ export class EUserTask {
 // DAO for one recording clip
 export class ERecording {
   parent: EUser;
-  path: string;  // Firestore document path
-  metadata: schema.ERecordingMetadata;
+  path: string; // Firestore document path
+  metadata: ERecordingMetadata;
 
-  constructor(parent: EUser, path: string, fsdata: schema.ERecordingData) {
+  constructor(parent: EUser, path: string, fsdata: ERecordingData) {
     this.parent = parent;
     this.path = path;
     this.metadata = JSON.parse(fsdata.metadata);
   }
 
-  async delete(user: EUser, usertask: EUserTask, tsTask: ETask, txn: Transaction): Promise<void> {
-    if (user !== this.parent || usertask.info.recordedTimestamp !== this.metadata.timestamp) {
-      throw new Error(`Unexpected wrong user/task update: ${user.euid}: ${this.metadata.timestamp}`);
+  async delete(
+    user: EUser,
+    usertask: EUserTask,
+    tsTask: ETask,
+    txn: Transaction,
+  ): Promise<void> {
+    if (
+      user !== this.parent ||
+      usertask.info.recordedTimestamp !== this.metadata.timestamp
+    ) {
+      throw new Error(
+        `Unexpected wrong user/task update: ${user.euid}: ${this.metadata.timestamp}`,
+      );
     }
     // The user's recording and task count should decline by one
-    this.parent.info.numCompletedTasks = Math.max(0, this.parent.info.numCompletedTasks - 1);
-    this.parent.info.numRecordings = Math.max(0, this.parent.info.numRecordings - 1);
+    this.parent.info.numCompletedTasks = Math.max(
+      0,
+      this.parent.info.numCompletedTasks - 1,
+    );
+    this.parent.info.numRecordings = Math.max(
+      0,
+      this.parent.info.numRecordings - 1,
+    );
     this.parent.update(txn);
 
     // The master task counter should decline by one
@@ -856,34 +1069,40 @@ export class ERecording {
     usertask.update(txn);
 
     // Delete the recording itself
-    txn.delete(this.parent.parent.firestore.doc(this.path));
+    txn.delete(doc(this.parent.parent.firestore, this.path));
   }
 }
 
 // DAO for one set of tasks
 export class ETaskSet {
   parent: EStorage;
-  path: string;  // Firestore document path
-  info: schema.ETaskSetInfo;
+  path: string; // Firestore document path
+  info: ETaskSetInfo;
 
-  constructor(parent: EStorage, path: string, fsdata: schema.ETaskSetData) {
-    this.parent = parent;  // EStorage system
+  constructor(parent: EStorage, path: string, fsdata: ETaskSetData) {
+    this.parent = parent; // EStorage system
     this.path = path;
     this.info = JSON.parse(fsdata.info);
   }
 
   // Returns all of this TaskSet's tasks.
   async listTasks(): Promise<ETask[]> {
-    const data = await this.parent.getTasksSubcollection(this.info.id).orderBy('order').get();
-    return data.docs.map(doc => new ETask(this, doc.ref.path, doc.data() as schema.ETaskData));
+    const data = await getDocs(
+      query(this.parent.getTasksSubcollection(this.info.id), orderBy("order")),
+    );
+
+    return data.docs.map(
+      (doc) => new ETask(this, doc.ref.path, doc.data() as ETaskData),
+    );
   }
 
   // Reads one task via a transaction.
   async loadTask(txn: Transaction, taskId: string): Promise<ETask> {
     const path = schema.taskPath(this.info.id, taskId);
-    const existingTaskData = await txn.get(this.parent.firestore.doc(path));
-    if (existingTaskData.exists) {
-      return new ETask(this, path, existingTaskData.data() as schema.ETaskData);
+    const existingTaskData = await txn.get(doc(this.parent.firestore, path));
+
+    if (existingTaskData.exists()) {
+      return new ETask(this, path, existingTaskData.data() as ETaskData);
     } else {
       throw new NotFoundError(`No such task: ${this.info.id}/${taskId}`);
     }
@@ -899,39 +1118,53 @@ export class ETaskSet {
 
     if (taskIds.length <= 10) {
       // Firestore IN clauses only support 10 terms
-      const q = this.parent.getTasksSubcollection(this.info.id).where(FieldPath.documentId(), 'in', taskIds);
-      const data = await q.get();
-      return ordered(data.docs.map(doc => new ETask(this, doc.ref.path, doc.data() as schema.ETaskData)));
+      const q = query(
+        this.parent.getTasksSubcollection(this.info.id),
+        where(documentId(), "in", taskIds),
+      );
 
+      const data = await getDocs(q);
+
+      return ordered(
+        data.docs.map(
+          (doc) => new ETask(this, doc.ref.path, doc.data() as ETaskData),
+        ),
+      );
     } else {
       // Scan the whole collection, assuming this is more efficient than many lookups.
       const set = new Set(taskIds);
       const result: ETask[] = [];
-      const data = await this.parent.getTasksSubcollection(this.info.id).get();
-      data.docs.map(doc => {
+      const data = await getDocs(
+        this.parent.getTasksSubcollection(this.info.id),
+      );
+      data.docs.map((doc) => {
         if (set.has(doc.id)) {
-          result.push(new ETask(this, doc.ref.path, doc.data() as schema.ETaskData));
+          result.push(new ETask(this, doc.ref.path, doc.data() as ETaskData));
         }
       });
+
       return ordered(result);
     }
   }
 
   // Returns a randomly selected subset of this set's Tasks.
   async sampleTasks(sampleSize: number): Promise<ETask[]> {
-    const data = await this.parent.getTasksSubcollection(this.info.id).get();
+    const data = await getDocs(this.parent.getTasksSubcollection(this.info.id));
     const docs = [...data.docs];
     shuffle(docs);
     if (sampleSize < docs.length) {
-      docs.splice(sampleSize, docs.length - sampleSize);  // erase the later elements
+      docs.splice(sampleSize, docs.length - sampleSize); // erase the later elements
     }
-    return docs.map(doc => new ETask(this, doc.ref.path, doc.data() as schema.ETaskData));
+
+    return docs.map(
+      (doc) => new ETask(this, doc.ref.path, doc.data() as ETaskData),
+    );
   }
 
   // Adds or removes TaskSet rules in memory; be sure to update(txn) after this.
-  changeRules(addrules: schema.EAssignmentRule[], delrules: number[]): void {
+  changeRules(addrules: EAssignmentRule[], delrules: number[]): void {
     if (addrules.length === 0 && delrules.length === 0) {
-      throw new ParamError('No rule changes specified');
+      throw new ParamError("No rule changes specified");
     }
 
     // Splice out unwanted rules from the list
@@ -951,11 +1184,14 @@ export class ETaskSet {
   // Commits changes to TaskSet properties
   async update(txn: Transaction) {
     const info = JSON.stringify(this.info);
-    txn.update(this.parent.firestore.doc(this.path), {info});
+    txn.update(doc(this.parent.firestore, this.path), { info });
   }
 
   // Tries to create new unique tasks. Duplicate prompts are prohibited.
-  async addTasks(taskType: schema.TaskType, tasks: Array<[number, string]>): Promise<ETask[]> {
+  async addTasks(
+    taskType: TaskType,
+    tasks: Array<[number, string]>,
+  ): Promise<ETask[]> {
     const tasksSubcollection = this.parent.getTasksSubcollection(this.info.id);
     for (const [order, prompt] of tasks) {
       if (!prompt || !order || isNaN(order)) {
@@ -966,26 +1202,36 @@ export class ETaskSet {
     const results: ETask[] = [];
     let pos = 0;
     while (pos < tasks.length) {
-      const r = await this.parent.firestore.runTransaction(async txn => {
-        const result: ETask[] = [];
-        for (let i = pos; i < tasks.length && i < pos + firebaseconfig.MAX_TASK_BATCH_SIZE; i++) {
-          // TODO: prevent duplicate prompts
-          const [order, prompt] = tasks[i];
-          const newDoc = tasksSubcollection.doc();
-          const info: schema.ETaskInfo = {
-            id: newDoc.id,
-            order,
-            taskType,
-            prompt,
-            creationTimestamp: Date.now(),
-            numRecordings: 0
-          };
-          const fsdata: schema.ETaskData = {order, info: JSON.stringify(info)};
-          txn.set(newDoc, fsdata);
-          result.push(new ETask(this, newDoc.path, fsdata));
-        }
-        return result;
-      });
+      const r = await runTransaction(
+        this.parent.firestore,
+        async (txn: Transaction) => {
+          const result: ETask[] = [];
+          for (
+            let i = pos;
+            i < tasks.length && i < pos + firebaseconfig.MAX_TASK_BATCH_SIZE;
+            i++
+          ) {
+            // TODO: prevent duplicate prompts
+            const [order, prompt] = tasks[i];
+            const newDoc = doc(tasksSubcollection);
+            const info: ETaskInfo = {
+              id: newDoc.id,
+              order,
+              taskType,
+              prompt,
+              creationTimestamp: Date.now(),
+              numRecordings: 0,
+            };
+            const fsdata: ETaskData = {
+              order,
+              info: JSON.stringify(info),
+            };
+            txn.set(newDoc, fsdata);
+            result.push(new ETask(this, newDoc.path, fsdata));
+          }
+          return result;
+        },
+      );
       pos += r.length;
       results.push.apply(results, r);
     }
@@ -993,7 +1239,7 @@ export class ETaskSet {
   }
 
   // Assigns a selection of tasks from this task set to one user
-  async assignTasks(spec: schema.EAssignmentRule, euid: string): Promise<EUser> {
+  async assignTasks(spec: EAssignmentRule, euid: string): Promise<EUser> {
     // Load the selected tasks matched by the rule
     let tasks: ETask[];
     if (spec.allTasks) {
@@ -1003,19 +1249,19 @@ export class ETaskSet {
     } else if (spec.sample > 0) {
       tasks = await this.sampleTasks(spec.sample);
     } else {
-      throw new ParamError('Malformed selection rule');
+      throw new ParamError("Malformed selection rule");
     }
 
     if (tasks.length === 0) {
-      throw new NotFoundError('No tasks to assign');
+      throw new NotFoundError("No tasks to assign");
     }
 
     // Write the tasks to the user in a transaction
     let u: EUser;
     let t: ETaskSet;
     for (const taskBatch of toBatches(tasks, 450)) {
-      [u, t] = await this.parent.run(async txn => {
-        const ts = await this.parent.requireTaskSet(this.info.id, txn);  // transactionally reload
+      [u, t] = await this.parent.run(async (txn) => {
+        const ts = await this.parent.requireTaskSet(this.info.id, txn); // transactionally reload
         const user = await this.parent.loadUser(euid, txn);
         await user.assignTasks(txn, taskBatch, ts);
         return [user, ts];
@@ -1032,10 +1278,10 @@ export class ETaskSet {
 // DAO for one Task on a TaskSet
 export class ETask {
   parent: ETaskSet;
-  path: string;  // Forestore document path
-  info: schema.ETaskInfo;
+  path: string; // Forestore document path
+  info: ETaskInfo;
 
-  constructor(parent: ETaskSet, path: string, fsdata: schema.ETaskData) {
+  constructor(parent: ETaskSet, path: string, fsdata: ETaskData) {
     this.parent = parent;
     this.path = path;
     this.info = JSON.parse(fsdata.info);
@@ -1043,18 +1289,20 @@ export class ETask {
 
   // Updates this Task with any changes carried by this DAO, such as counter changes.
   update(txn: Transaction) {
-    txn.update(this.parent.parent.firestore.doc(this.path), {
-      info: JSON.stringify(this.info)
+    txn.update(doc(this.parent.parent.firestore, this.path), {
+      info: JSON.stringify(this.info),
     });
   }
 
   // Adds a graphical image to the task, and stores it. Any previous image is replaced.
   async addImage(contents: Buffer, txn: Transaction): Promise<void> {
     // Store the image data. TODO: re-render this image so it's scrubbed for metadata / malice / format
-    await this.parent.parent.getImageFile(this.parent.info.id, this.info.id).save(contents);
+    await this.parent.parent
+      .getImageFile(this.parent.info.id, this.info.id)
+      .save(contents);
 
     // Once the contents are written, add the metadata. We only support JPG for now.
-    this.info.imageType = 'image/jpeg';
+    this.info.imageType = "image/jpeg";
     this.update(txn);
   }
 }
@@ -1062,17 +1310,17 @@ export class ETask {
 // DAO for one Consent
 export class EConsent {
   parent: EStorage;
-  path: string;  // Firestore document path
-  info: schema.EConsentInfo;
+  path: string; // Firestore document path
+  info: EConsentInfo;
 
-  constructor(parent: EStorage, path: string, fsdata: schema.EConsentData) {
-    this.parent = parent;  // EStorage system
+  constructor(parent: EStorage, path: string, fsdata: EConsentData) {
+    this.parent = parent; // EStorage system
     this.path = path;
     this.info = JSON.parse(fsdata.info);
   }
 
   // Returns a variation of the contained consent info that has only the live version and no others.
-  infoWithOnlyLiveVersion(now: number): schema.EConsentInfo {
+  infoWithOnlyLiveVersion(now: number): EConsentInfo {
     const live = this.getLiveVersion(now)!;
     const info = clone(this.info);
     info.versions = [clone(live)];
@@ -1080,10 +1328,13 @@ export class EConsent {
   }
 
   // Returns the latest live version, unless no version is live yet.
-  getLiveVersion(now: number): schema.EConsentVersion|undefined {
-    let latest: schema.EConsentVersion|undefined;
+  getLiveVersion(now: number): EConsentVersion | undefined {
+    let latest: EConsentVersion | undefined;
     for (const version of this.info.versions) {
-      if (version.liveTimestamp <= now && (!latest || latest.version < version.version)) {
+      if (
+        version.liveTimestamp <= now &&
+        (!latest || latest.version < version.version)
+      ) {
         latest = version;
       }
     }
@@ -1104,24 +1355,34 @@ export class EConsent {
   changeUserCount(version: number, incr: number) {
     const idx = this.getVersionIndex(version);
     if (idx === -1) {
-      throw new NotFoundError(`No such consent version: ${this.info.id}-${version}`);
+      throw new NotFoundError(
+        `No such consent version: ${this.info.id}-${version}`,
+      );
     }
     this.info.versions[idx].numUsers += incr;
   }
 
   // Adds a new version, stores the text in GCS, and commits it
-  async createVersion(description: string, liveTimestamp: number, contents: Buffer, txn: Transaction): Promise<void> {
+  async createVersion(
+    description: string,
+    liveTimestamp: number,
+    contents: Buffer,
+    txn: Transaction,
+  ): Promise<void> {
     // Find the lowest unused version number
-    const version = this.info.versions.reduce((n, v) => Math.max(n, v.version), 0) + 1;
-    
+    const version =
+      this.info.versions.reduce((n, v) => Math.max(n, v.version), 0) + 1;
+
     // Store the consent text; these are immutable
     await this.parent.getConsentFile(this.info.id, version).save(contents);
 
     // Once the contents are written, add the metadata
     this.info.versions.push({
-      version, liveTimestamp, description,
+      version,
+      liveTimestamp,
+      description,
       creationTimestamp: Date.now(),
-      numUsers: 0
+      numUsers: 0,
     });
     this.update(txn);
   }
@@ -1130,12 +1391,16 @@ export class EConsent {
   async deleteVersion(version: number, txn: Transaction): Promise<void> {
     const idx = this.getVersionIndex(version);
     if (idx === -1) {
-      throw new NotFoundError(`No such consent version: ${this.info.id}-${version}`);
+      throw new NotFoundError(
+        `No such consent version: ${this.info.id}-${version}`,
+      );
     }
 
     const versionData = this.info.versions[idx];
     if (versionData.numUsers > 0) {
-      throw new ParamError(`Consent version is in use and cannot be removed: ${this.info.id}-${version}`);
+      throw new ParamError(
+        `Consent version is in use and cannot be removed: ${this.info.id}-${version}`,
+      );
     }
     this.info.versions.splice(idx, 1);
     this.update(txn);
@@ -1147,6 +1412,6 @@ export class EConsent {
   // Commits changes to Consent properties. Don't use this for version creation or deletion.
   update(txn: Transaction) {
     const info = JSON.stringify(this.info);
-    txn.update(this.parent.firestore.doc(this.path), {info});
+    txn.update(doc(this.parent.firestore, this.path), { info });
   }
 }
